+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Changelog
 * Add new method :meth:`pybed.BedFrame.from_regions`.
 * Update :meth:`pyvcf.VcfFrame.from_file` method to accept BED data to specify regions of interest.
 * Update :command:`vcf-slice` command to run significantly faster by allowing random access.
+* Add new method :meth:`common.sort_regions`.
 
 0.25.0 (2021-10-09)
 -------------------